changed urls to different file
