+++ config/pageURL.ts
@@ -1,15 +1,4 @@
-export const testConfig = {
-  baseUrl: process.env.BASE_URL,
-  userName: process.env.USERNAME,
-  password: process.env.PASSWORD,
-  empty_credentials: "Please provide USERNAME and PASSWORD",
-  empty_url: "Please provide the BASE_URL",
-  startPageTitle:
-    " Apply to register as a Companies House authorised agent - GOV.UK",
-  nameTitle:
-    "What is your name? - Apply to register as a Companies House authorised agent - GOV.UK",
-  dobTitle:
-    "What is your date of Birth? - Apply to register as a Companies House authorised agent - GOV.UK",
+export const pageURL = {
   typeOfBusiness: "/what-business-type",
   soleTrader: {
     name: "/sole-trader/what-is-your-name",
