+++ config/testConfig.ts
@@ -1,0 +1,13 @@
+export const testConfig = {
+  baseUrl: process.env.BASE_URL,
+  userName: process.env.USERNAME,
+  password: process.env.PASSWORD,
+  empty_credentials: "Please provide USERNAME and PASSWORD",
+  empty_url: "Please provide the BASE_URL",
+  startPageTitle:
+    " Apply to register as a Companies House authorised agent - GOV.UK",
+  nameTitle:
+    "What is your name? - Apply to register as a Companies House authorised agent - GOV.UK",
+  dobTitle:
+    "What is your date of Birth? - Apply to register as a Companies House authorised agent - GOV.UK",
+};
