+++ tests/accessibility/soleTrader/soleTraderscreens.test.ts
@@ -1,5 +1,7 @@
 import { Page, expect, test } from "@playwright/test";
-import { testConfig } from "../../../testConfig";
+import { testConfig } from "../../../config/testConfig";
+import { pageURL } from "../../../config/pageURL";
+
 import { accessibilityScan } from "../../../utils/accessibilityScan";
 import { globalSetUp } from "../../../setUp/globalSetup";
 
@@ -17,7 +19,7 @@ test("Accessibility check for Sole-Trader-Name screen @accessibility", async ({
   await accessibilityContext.checkWcagCompliance(
     page,
 
-    testConfig.baseUrl + testConfig.soleTrader.name,
+    testConfig.baseUrl + pageURL.soleTrader.name,
 
     testInfo
   );
@@ -30,7 +32,7 @@ test("Accessibility check for Sole Trader Date of Birth screen @accessibility", 
 
   await accessibilityContext.checkWcagCompliance(
     page,
-    testConfig.baseUrl + testConfig.soleTrader.dateOfBirth,
+    testConfig.baseUrl + pageURL.soleTrader.dateOfBirth,
     testInfo
   );
 });
@@ -42,7 +44,7 @@ test("Accessibility check for Sole Trader Nationality screen @accessibility", as
 
   await accessibilityContext.checkWcagCompliance(
     page,
-    testConfig.baseUrl + testConfig.soleTrader.nationality,
+    testConfig.baseUrl + pageURL.soleTrader.nationality,
     testInfo
   );
 });
@@ -54,7 +56,7 @@ test("Accessibility check for Sole Trader Where do you Live screen @accessibilit
 
   await accessibilityContext.checkWcagCompliance(
     page,
-    testConfig.baseUrl + testConfig.soleTrader.whereDoYouLive,
+    testConfig.baseUrl + pageURL.soleTrader.whereDoYouLive,
     testInfo
   );
 });
@@ -66,7 +68,7 @@ test("Accessibility check for Sole Trader Correspondence Address Auto-lookup scr
 
   await accessibilityContext.checkWcagCompliance(
     page,
-    testConfig.baseUrl + testConfig.soleTrader.addressAutoLookUp,
+    testConfig.baseUrl + pageURL.soleTrader.addressAutoLookUp,
     testInfo
   );
 });
@@ -78,7 +80,7 @@ test("Accessibility check for Sole Trader Correspondence Address Select Address 
 
   await accessibilityContext.checkWcagCompliance(
     page,
-    testConfig.baseUrl + testConfig.soleTrader.addressSelect,
+    testConfig.baseUrl + pageURL.soleTrader.addressSelect,
     testInfo
   );
 });
@@ -90,7 +92,7 @@ test("Accessibility check for Sole Trader Correspondence Address Manual Entry sc
 
   await accessibilityContext.checkWcagCompliance(
     page,
-    testConfig.baseUrl + testConfig.soleTrader.addressManualEntry,
+    testConfig.baseUrl + pageURL.soleTrader.addressManualEntry,
     testInfo
   );
 });
@@ -114,21 +116,19 @@ test("Accessibility check for Sole Trader Which sector screen @accessibility", a
 
   await accessibilityContext.checkWcagCompliance(
     page,
-    testConfig.baseUrl + testConfig.soleTrader.whichSector,
+    testConfig.baseUrl + pageURL.soleTrader.whichSector,
     testInfo
   );
-
-  });
-
-  test("Accessibility check for Sole Trader Which sector-Other screen @accessibility", async ({
-    page,
-  }, testInfo) => {
-    const accessibilityContext = new accessibilityScan();
-  
-    await accessibilityContext.checkWcagCompliance(
-      page,
-      testConfig.baseUrl + testConfig.soleTrader.whichSectorOther,
-      testInfo
-    );
 });
 
+test("Accessibility check for Sole Trader Which sector-Other screen @accessibility", async ({
+  page,
+}, testInfo) => {
+  const accessibilityContext = new accessibilityScan();
+
+  await accessibilityContext.checkWcagCompliance(
+    page,
+    testConfig.baseUrl + pageURL.soleTrader.whichSectorOther,
+    testInfo
+  );
+});
