+++ tests/functional/userJourney/soleTrader.test.ts
@@ -1,7 +1,8 @@
 import { expect, test } from "@playwright/test";
 import { dobPage } from "../../../pages/soleTrader/dobPage";
 import { namePage } from "../../../pages/soleTrader/namePage";
-import { testConfig } from "../../../testConfig";
+import { testConfig } from "../../../config/testConfig";
+import { pageURL } from "../../../config/pageURL";
 import { userInput } from "../../../testdata/userInput";
 import { userActions } from "../../../utils/userActions";
 import { globalSetUp } from "../../../setUp/globalSetup";
@@ -16,7 +17,7 @@ test("Verify Sole Trader can register as an ACSP, @smoke", async ({ page }) => {
   const namePageContext = new namePage(page);
   const userActionsContext = new userActions(page);
   const dobPageContext = new dobPage(page);
-  await page.goto(testConfig.baseUrl + testConfig.soleTrader.name);
+  await page.goto(testConfig.baseUrl + pageURL.soleTrader.name);
   await expect(page).toHaveTitle(testConfig.nameTitle);
   await expect(namePageContext.firstName).toBeVisible();
   await namePageContext.firstName.fill(userInput.firstName);
